backend/app/requirements.txt
fastapi 
uvicorn
pymysql
sqlalchemy
python-jose
pydantic
python-dotenv
passlib